🎨 (listerProjetsÀNotifier): query déplacée dans infra/sequlize/queries
