+++ src/modules/notificationCandidats/index.ts
@@ -2,5 +2,4 @@ export * from './eventHandlers'
 export * from './events'
 export * from './helpers'
 export * from './queries'
-export * from './useCases'
 export * from './CandidateNotification'
