+++ src/modules/notificationCandidats/queries/index.ts
@@ -1,1 +1,2 @@
 export * from './GetPendingCandidateInvitations'
+export * from './ListerProjetsÀNotifier'
