--- src/modules/notificationCandidats/useCases/index.ts
@@ -1,1 +1,0 @@
-export * from './listerProjetsÀNotifier'
